pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-kb"
version = "1.0.0"
description = "轻量级 RAG 知识库管理工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "markitdown",
    "chonkie",
    "qdrant-client",
    "rank-bm25",
    "pyyaml",
    "requests",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.scripts]
rag = "rag_kb.cli:run_cli"
rag-server = "rag_kb.server:main"

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src/rag_kb"]